requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.117.1
graphql-core==3.2.6
greenlet==3.2.4
h11==0.16.0
idna==3.10
lia-web==0.2.3
packaging==25.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
strawberry-graphql==0.282.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
